Added ability to specify team name
This will annotate all created resources with the team name, this is
required in order for alerting to function property as we have to know
who to alert.
Made adjustments to various tests to include team name as part of tests
where it made sense.
Made new test for the "createObjectMeta" helper function.