Add more tests for creation of Prometheus Alerts
Added smoke-screen test in api that actually posts a manifest with a
rule specified and checks whether or not the api responds with the
appropriate deployment status message.

Also added test for "addRulesToConfigMap" that adds a single rule to a
config map, parses the config map yaml back into a rule and
verifies they are equal.