When deploy to app namespace, delete old app
Every time an application is deployed, attempt to forward traffic from
the old service to the newly created one in app-namespace.

If it's the first time the app is deployed, it is important that we wait
until the newly created pod is Ready before the send traffic to it, then
delete the old deployment.

Put clientholder in own file, as it's used many places.