Allow service spec updates.
In order to allow switching between application-namespaces and the old
behaviour we need to support updating service definitions.  This is a
little bit tricky because many of the fields in the service are
immutable.  In order for this to work, the method fillServiceSpec was
created, which takes an existing service spec and changes only mutable
values.