Use separate web proxy settings for all pods (#122)
This patch changes the source of pods' web proxy configuration from `HTTP_PROXY` and `NO_PROXY` to `NAIS_POD_HTTP_PROXY` and `NAIS_POD_NO_PROXY`, respectively. The proxy settings are injected automatically without the need to specify this need in the manifest.

Additionally, a `JAVA_PROXY_OPTIONS` environment variable is added to all pods, to ease the pain of having to set up this separately in Java applications. We have several different implementations of this at NAV, and this approach will provide users with platform-provided values so that they can focus on delivering applications instead of tinkering with infrastructure details.